+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/animation/viewport/MapViewportState.kt
@@ -247,7 +247,7 @@ public class MapViewportState(
    */
   @MapboxExperimental
   public fun transitionToFollowPuckState(
-    followPuckViewportStateOptions: FollowPuckViewportStateOptions,
+    followPuckViewportStateOptions: FollowPuckViewportStateOptions = FollowPuckViewportStateOptions.Builder().build(),
     defaultTransitionOptions: DefaultViewportTransitionOptions = DefaultViewportTransitionOptions.Builder()
       .build(),
     completionListener: CompletionListener? = null